De-flake relay integration suite: shared IntegrationHttpSupport helper + SEPARATE_THREAD timeouts
Extract the duplicated socket-HTTP request/response helper (the readLine()-racing-soTimeout source of the SocketTimeoutException flakes) into a shared core/testfixtures IntegrationHttpSupport, and convert ClaudeFullFlowEndToEndTest, OAuthEndToEndTest, TunnelMcpIntegrationTest, EndToEndSessionTest, and MultiClientConcurrencyTest to it. Robust read backed by a SEPARATE_THREAD hard @Timeout ceiling rather than racing per-read soTimeout. Test infra only. #501
