+++ core/tunnel/src/jvmTest/kotlin/com/rousecontext/tunnel/integration/MultiClientConcurrencyTest.kt
@@ -20,9 +20,7 @@ import io.modelcontextprotocol.kotlin.sdk.server.Server
 import io.modelcontextprotocol.kotlin.sdk.types.CallToolResult
 import io.modelcontextprotocol.kotlin.sdk.types.TextContent
 import io.modelcontextprotocol.kotlin.sdk.types.ToolSchema
-import java.io.BufferedReader
 import java.io.File
-import java.io.InputStreamReader
 import java.security.KeyStore
 import java.security.MessageDigest
 import java.security.SecureRandom
@@ -72,7 +70,13 @@ import org.junit.jupiter.api.Timeout
 @Suppress("LargeClass")
 @Tag("integration")
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@Timeout(value = 180, unit = TimeUnit.SECONDS)
+// SEPARATE_THREAD so the class ceiling also covers the @BeforeAll relay setup
+// and fails fast on a blocked read (the per-method ceiling below is tighter).
+@Timeout(
+    value = 180,
+    unit = TimeUnit.SECONDS,
+    threadMode = Timeout.ThreadMode.SEPARATE_THREAD
+)
 class MultiClientConcurrencyTest {
 
     companion object {
@@ -643,7 +647,7 @@ class MultiClientConcurrencyTest {
             "127.0.0.1",
             relayPort
         ) as SSLSocket
-        socket.soTimeout = 30_000
+        IntegrationHttpSupport.applyReadTimeout(socket)
         val params = socket.sslParameters
         params.serverNames = listOf(javax.net.ssl.SNIHostName(INTEGRATION_HOST))
         socket.sslParameters = params
@@ -655,22 +659,11 @@ class MultiClientConcurrencyTest {
     // HTTP helpers
     // =====================================================================
 
-    private data class HttpResponse(
-        val statusCode: Int,
-        val headers: Map<String, String>,
-        val body: String
-    )
-
     private var requestIdCounter = 100
 
     private fun nextId(): Int = ++requestIdCounter
 
-    @Suppress(
-        "LongParameterList",
-        "LongMethod",
-        "CyclomaticComplexMethod",
-        "LoopWithTooManyJumpStatements"
-    )
+    @Suppress("LongParameterList")
     private fun httpRequest(
         input: java.io.InputStream,
         output: java.io.OutputStream,
@@ -680,13 +673,13 @@ class MultiClientConcurrencyTest {
         bearerToken: String? = null,
         contentType: String = "application/json",
         mcpSessionId: String? = null
-    ): HttpResponse {
+    ): IntegrationHttpSupport.HttpResponse {
         val sb = StringBuilder()
         sb.append("$method $path HTTP/1.1\r\n")
         sb.append("Host: $INTEGRATION_HOST\r\n")
 
-        if (body != null) {
-            val bodyBytes = body.toByteArray(Charsets.UTF_8)
+        val bodyBytes = body?.toByteArray(Charsets.UTF_8)
+        if (bodyBytes != null) {
             sb.append("Content-Type: $contentType\r\n")
             sb.append("Content-Length: ${bodyBytes.size}\r\n")
         }
@@ -699,80 +692,7 @@ class MultiClientConcurrencyTest {
         }
         sb.append("\r\n")
 
-        output.write(sb.toString().toByteArray(Charsets.UTF_8))
-        if (body != null) {
-            output.write(body.toByteArray(Charsets.UTF_8))
-        }
-        output.flush()
-
-        val reader = BufferedReader(InputStreamReader(input, Charsets.UTF_8))
-        val statusLine = reader.readLine() ?: error("No status line received")
-        assertTrue(
-            statusLine.startsWith("HTTP/1.1"),
-            "Expected HTTP response, got: $statusLine"
-        )
-        val statusCode = statusLine.split(" ")[1].toInt()
-
-        val headers = mutableMapOf<String, String>()
-        var contentLength = -1
-        var chunked = false
-        while (true) {
-            val line = reader.readLine() ?: break
-            if (line.isEmpty()) break
-            val lower = line.lowercase()
-            val colonIdx = line.indexOf(':')
-            if (colonIdx > 0) {
-                headers[line.substring(0, colonIdx).lowercase()] =
-                    line.substring(colonIdx + 1).trim()
-            }
-            if (lower.startsWith("content-length:")) {
-                contentLength = lower.substringAfter(":").trim().toInt()
-            }
-            if (lower.startsWith("transfer-encoding:") && lower.contains("chunked")) {
-                chunked = true
-            }
-        }
-
-        val responseBody = when {
-            contentLength > 0 -> {
-                val buf = CharArray(contentLength)
-                var read = 0
-                while (read < contentLength) {
-                    val n = reader.read(buf, read, contentLength - read)
-                    if (n == -1) break
-                    read += n
-                }
-                String(buf, 0, read)
-            }
-            chunked -> readChunkedBody(reader)
-            contentLength == 0 -> ""
-            else -> ""
-        }
-
-        return HttpResponse(statusCode, headers, responseBody)
-    }
-
-    @Suppress("LoopWithTooManyJumpStatements")
-    private fun readChunkedBody(reader: BufferedReader): String {
-        val sb = StringBuilder()
-        while (true) {
-            val sizeLine = reader.readLine() ?: break
-            val chunkSize = sizeLine.trim().toInt(16)
-            if (chunkSize == 0) {
-                reader.readLine()
-                break
-            }
-            val buf = CharArray(chunkSize)
-            var read = 0
-            while (read < chunkSize) {
-                val n = reader.read(buf, read, chunkSize - read)
-                if (n == -1) break
-                read += n
-            }
-            sb.append(buf, 0, read)
-            reader.readLine()
-        }
-        return sb.toString()
+        return IntegrationHttpSupport.exchange(input, output, sb.toString(), bodyBytes)
     }
 
     // =====================================================================
